==> 2_peak_detection.Rmd <==
---
title: "XCMS processing"
author: "Mar Garcia-Aloy"
output: 
  BiocStyle::html_document:
    toc: true
    number_sections: false
    toc_float: true
---

```{r startpoint, include = FALSE}
startpoint <- Sys.time()
```

In this document I'm going to import DDA data on study samples in order to 
select those peaks that have at least 1 MS2 spectra.  
Fragmentation spectra has obtained using data dependent acquisition (DDA) 
method. In this method the top N most intense m/z values from a MS1 scan are 
selected for fragmentation in the next N scans before the cycle starts again.  
The acquired data contains both MS1 and MS2 spectra.

https://bioconductor.org/packages/release/bioc/vignettes/xcms/inst/doc/xcms-lcms-ms.html#LC-MSMS_data_analysis_with_xcms


# Libraries

```{r libraries, message = FALSE, warning = FALSE}
library(xcms)
library(MsFeatures)
library(MetaboCoreUtils)
```


# Data import

```{r import, eval = !file.exists("data/tmp/dda_data_tissues_POS.RData")}
for(s in c("tissues", "maturation")){
  for(p in c("POS", "NEG")){
    fln <- list.files(paste0("data/", s, "/", p, "_DDA_mzML/"))
    fln <- fln[!grepl("MIX|QC|entire|STD|solv", fln)]
    mtd <- gsub("Rep_", "rep", fln)
    mtd <- gsub("Pt_", "pt", mtd)
    mtd <- gsub("lipidgrape_tissues", "lipidgrape", mtd)
    if(s == "tissues"){
      mtd <- gsub("pt11_", "", mtd)
    }
    mtd <- strsplit(gsub(".mzML", "", mtd), "_")
    mtd <- data.frame(do.call(rbind, mtd))
    colnames(mtd) <- c("order", "project", "class", "rep", "p", "mode")
    injections <- data.frame(cbind(fln, mtd))
    
    # Read the data
    dda_data <- readMSData(
      paste0("data/", s, "/", p, "_DDA_mzML/" , injections$fln),
      pdata = as(AnnotatedDataFrame(injections),
                 "NAnnotatedDataFrame"), 
      mode = "onDisk")
    
    assign(paste("dda_data", s, p, sep = "_"), dda_data)
    save(dda_data, file = paste0("data/tmp/dda_data_", s, "_", p, ".RData"))
  }
}
rm(injections, fln, mtd, dda_data) #s, p
```

```{r load, eval = file.exists("data/tmp/dda_data_tissues_POS.RData")}
for(s in c("tissues", "maturation")){
  for(p in c("POS", "NEG")){
    load(paste0("data/tmp/dda_data_", s, "_", p, ".RData"))
    assign(paste("dda_data", s, p, sep = "_"), dda_data)
  }
}
```


# Peak detection & grouping

Here we perform the chromatographic peak detection on the MS level 1 data, 
followed by peak grouping.

```{r peak-detection}
for(s in c("tissues", "maturation")){
  for(p in c("POS", "NEG")){
    dda_xdata <- get(paste("dda_data", s, p, sep = "_"))
    if(p == "NEG"){
      int <- 1e4
    } else if(p == "POS"){
      int <- 1e5
    }
    dda_xdata <- findChromPeaks(
      dda_xdata, 
      param = CentWaveParam(ppm = 20,
                            peakwidth = c(5, 40),
                            prefilter = c(5, int),
                            snthresh = 5,
                            noise = 100,
                            mzdiff = 0.001,
                            integrate = 2)
    )
    register(SerialParam())
    dda_xdata <- refineChromPeaks(
      dda_xdata, 
      param = FilterIntensityParam(nValues = 5, threshold = int)
    )
    dda_xdata <- groupChromPeaks(
      dda_xdata, 
      param = PeakDensityParam(sampleGroups = dda_xdata$class,
                               minFraction = 1,
                               binSize = 0.02,
                               bw = 3)
    )
    chromPeaks(dda_xdata)
    featureDefinitions(dda_xdata)
    assign(paste("dda_xdata", s, p, sep = "_"), dda_xdata)
  }
}
rm(int, dda_xdata)
```


# MS1 and MS2 data integration

The advantage of LC-MS/MS data is that (MS1) ions are fragmented and the 
corresponding MS2 spectra measured. Thus, for some of the ions (identified as 
MS1 chromatographic peaks) MS2 spectra are available.   
Below we firstly extract the MS2 spectra to the object `dda_spectra` and the 
features detected in the MS1 data to the object `ft`.

```{r ms2-integration1}
for(s in c("tissues", "maturation")){
  for(p in c("POS", "NEG")){
    dda_xdata <- get(paste("dda_xdata", s, p, sep = "_"))
    ft <- as.data.frame(featureDefinitions(dda_xdata))
    assign(paste("ft", s, p, sep = "_"), ft)
  }
}
rm(ft, dda_xdata)
```

For each feature we extract all MS2 spectra that were associated with the 
corresponding feature using the ID of the corresponding chromatographic 
peaks. We annotate how many MS2 spectra do we have for each feature in order
to keep only those features with at least 1 MS2 spectra.

```{r ms2-integration2}
for(s in c("tissues", "maturation")){
  for(p in c("POS", "NEG")){
    ft <- get(paste("ft", s, p, sep = "_"))
    dda_xdata <- get(paste("dda_xdata", s, p, sep = "_"))
    dda_spectra <- chromPeakSpectra(
      dda_xdata, msLevel = 2L, return.type = "Spectra")
    ft$n_MS2 <- NA
    for(i in seq(nrow((ft)))){
      i_id <- rownames(chromPeaks(dda_xdata)[ft[i, "peakidx"][[1]],])
      i_spectra <- dda_spectra[dda_spectra$peak_id %in% i_id]
      ft$n_MS2[i] <- length(i_spectra)
    }
    assign(paste("ft", s, p, sep = "_"), ft)
    assign(paste("dda_spectra", s, p, sep = "_"), dda_spectra)
  }
}

rm(ft, dda_spectra, dda_xdata, i, i_id, i_spectra)
```


# Isotopologues

Finally we want to detect isotopes using the function `isotopologues`.   
This function identifies isotopes in a given MS spectrum. However, our data 
is LC-MS. Therefore, first of all we group features according to retention 
time (http://sneumann.github.io/xcms/articles/LC-MS-feature-grouping.html), 
in order to have something similar to a pseudo-MS spectrum, and then within 
each feature group we apply the function.  

```{r isotopes}
for(s in c("tissues", "maturation")){
  for(p in c("POS", "NEG")){
    dda_xdata <- get(paste("dda_xdata", s, p, sep = "_"))
    ft <- get(paste("ft", s, p, sep = "_"))
    
    # feature grouping:
    dda_xdata <- groupFeatures(dda_xdata, param = SimilarRtimeParam(3))
    ft <- cbind(ft, featureGroups(dda_xdata))
    colnames(ft)[ncol(ft)] <- "FG"
    
    # calculate mean-intensities of each feature:
    dt <- data.frame(t(featureValues(dda_xdata, value = "into")))
    all(colnames(dt) == rownames(ft))
    ft$into <- apply(dt, 2, mean, na.rm = TRUE)
    
    # identify isotopes:
    ft$isotope <- NA
    ft$is_isotope <- FALSE
    ft <- ft[order(ft$mzmed),]
    xi <- 0
    for(j in unique(ft$FG)){ # within FG "j"....
      tmp <- ft[ft$FG == j,]
      if(nrow(tmp) > 1){ # if there is >1feature in FG "j", apply "isotopologues"
        isos <- isotopologues(as.data.frame(tmp[,c("mzmed", "into")]), 
                              ppm = 10)
        if(length(isos) > 0){ # if there are isotopes detected, annotate them
          for(i in seq_along(isos)){
            z <- isos[[i]]
            xi <- xi + 1
            z <- which(rownames(ft) %in% rownames(tmp)[z])
            ft$isotope[z] <- paste0(
              "i", formatC(xi, width = 2, flag = "0")
            )
            ft$is_isotope[z[-1]] <- TRUE
          }
        }
      }
    }
    ft$dataset <- paste(s, p, sep = "_")
    assign(paste("dda_xdata", s, p, sep = "_"), dda_xdata)
    assign(paste("ft", s, p, sep = "_"), ft)
  }
}
rm(dda_xdata, ft, dt, xi, j, tmp, isos, i, z, s, p)
```


# Save data

```{r save}
save.image("data/RData.RData")
```


# Session information

```{r session}
Sys.time()-startpoint
devtools::session_info()
```
