==> 3_identification.Rmd <==
---
title: "Identification"
author: "Mar Garcia-Aloy"
output: 
  BiocStyle::html_document:
    toc: true
    number_sections: false
    toc_float: true
---

# Libraries

```{r libraries, message = FALSE, warning = FALSE}
library(MetaboAnnotation)
library(xcms)
library(CompoundDb)
library(MetaboCoreUtils)
```


# Data import

```{r import}
load("data/RData.RData")
load("data/ionsdb.RData")
load("data/colors.RData")
load("data/MS2_STDmix.RData")
db_file <- "data/CompDb_lipids_2.sqlite"
startpoint <- Sys.time()
```


# Annotation

```{r annotation}
unify_dpls <- function(pks){
  dpl <- unique(gsub("\\..*", "", rownames(pks)[grep(
    "\\.", rownames(pks))]))
  for(i in dpl){
    idx <- c(which(rownames(pks) == i), grep(
      paste0(i, "."), rownames(pks)))
    tmp <- pks[idx,]
    pks$target_compound_id[idx[1]] <- paste(
      pks$target_compound_id[idx], collapse = "-")
    pks$target_ion_adduct[idx[1]] <- paste(
      unique(pks$target_ion_adduct[idx]), collapse = "-")
    pks$target_name[idx[1]] <- paste(
      unique(pks$target_name[idx]), collapse = "-")
    pks <- pks[-idx[-1],]
  }
  return(pks)
}

ionsdb$ion_rt <- ionsdb$ion_rt*60
for(s in c("tissues", "maturation")){
  for(p in c("POS", "NEG")){
    ft <- get(paste("ft", s, p, sep = "_"))
    ft <- ft[ft$n_MS2 > 0,] # select features with at least 1  MS2
    ft <- ft[!ft$is_isotope,] # select non-isotopic features
    # get the ions from polarity "p":
    if(p == "POS"){
      tmp_ionsdb <- ionsdb[
        ionsdb$ion_adduct %in% adds[adds$positive, "name"],]
    } else if(p == "NEG"){
      tmp_ionsdb <- ionsdb[
        ionsdb$ion_adduct %in% adds[!adds$positive, "name"],]
    }
    
    # start matching based on mz & RT:
    param <- MzRtParam(ppm = 5, toleranceRt = 5)
    pks_match <- matchMz(ft, tmp_ionsdb, param, 
                         mzColname = c("mzmed", "ion_mz"),
                         rtColname = c("rtmed", "ion_rt"))
    pks <- as.data.frame(matchedData(pks_match))
    pks_rt <- pks[!is.na(pks$target_compound_id),]
    
    pks_rt <- unify_dpls(pks_rt) # unify multiple matches
    
    # for those un-annotated features, use more relaxed parameters:
    param <- MzRtParam(ppm = 15, toleranceRt = 15)
    pks_match <- matchMz(ft[!rownames(ft) %in% rownames(pks_rt),], 
                         tmp_ionsdb, param, 
                         mzColname = c("mzmed", "ion_mz"),
                         rtColname = c("rtmed", "ion_rt"))
    pks2 <- as.data.frame(matchedData(pks_match))
    pks2 <- pks2[!is.na(pks2$target_compound_id),]
    
    pks2 <- unify_dpls(pks2) # unify multiple matches
    colnames(pks2)[ncol(pks)-1] <- "ppm_error"
    pks2 <- rbind(pks_rt, pks2)
    
    # for those un-annotated features, match based only on mz:
    if(nrow(pks_rt) < nrow(ft)){
      param <- MzParam(ppm = 5)
      pks_match <- matchMz(ft[!rownames(ft) %in% rownames(pks2),], 
                           tmp_ionsdb, param, 
                           mzColname = c("mzmed", "ion_mz"))
      pks_mz <- as.data.frame(matchedData(pks_match))
      pks_mz <- unify_dpls(pks_mz) # unify multiple matches
      pks_mz$score_rt <- NA
      colnames(pks_mz)[ncol(pks_mz)-1] <- "ppm_error"
      pks <- rbind(pks2, pks_mz)
    }
    
    pks <- pks[order(rownames(pks)),]
    pks <- pks[,c("mzmed", "rtmed", "n_MS2", "FG", "into", "isotope", 
                  "is_isotope", "target_compound_id", "target_ion_adduct", 
                  "target_name", "target_class", "ppm_error", "score_rt")]
    pks$FT <- rownames(pks)
    pks$rtmed <- pks$rtmed/60
    pks$dataset <- paste(s, p, sep = "_")
    assign(paste("pks", s, p, sep = "_"), pks)
    pks <- pks[!is.na(pks$target_compound_id),]
    pks <- pks[!grepl("-", pks$target_compound_id),]
    pks <- pks[order(pks$target_compound_id, pks$rtmed),]
  }
}
rm(ft, tmp_ionsdb, param, pks_match, pks_rt, pks)
```


# Workflow

```{r workflow}
for(s in c("tissues", "maturation")){
  for(p in c("POS", "NEG")){
    dda_xdata <- get(paste("dda_xdata", s, p, sep = "_"))
    ft <- get(paste("ft", s, p, sep = "_"))
    pks <- get(paste("pks", s, p, sep = "_"))
    print(paste(s, p))
    print(paste("Detected features in MS1 data:", 
                nrow(featureDefinitions(dda_xdata))))
    print(paste0("Isotopes: ", sum(ft$is_isotope), "(",
                round((sum(ft$is_isotope)*100)/nrow(ft), 1), "%)"))
    print(paste("Non-isotopic MS1 features:", sum(!ft$is_isotope)))
    pks <- pks[!pks$is_isotope,]
    print(paste("Non-isotopic MS1 features with at least one MS2 spectra:", 
                nrow(pks)))
    pks <- pks[!is.na(pks$target_compound_id),]
    pks <- pks[!grepl("-", pks$target_compound_id),]
    print(paste("Annotated features:", nrow(pks)))
    print(paste("Unique compounds:", length(unique(pks$target_compound_id))))
    print("################################") 
  }
}
rm(dda_xdata, ft, pks)
```


# Table

```{r table}
pks <- rbind(pks_tissues_POS, pks_maturation_POS,
             pks_tissues_NEG, pks_maturation_NEG)
pks <- pks[!is.na(pks$score_rt) & !grepl("-", pks$target_compound_id),]
cmps <- compounds(CompDb(db_file), c(
  "compound_id", "name", "formula", "exactmass", "class", "std", "instd", 
  "C", "db"))
cmps <- cmps[!grepl("tail", cmps$name),]
cmps <- cmps[cmps$compound_id %in% pks$target_compound_id,]
cmps$RT <- NA
cmps$tissues <- ""
cmps$maturation <- ""
for(i in seq(nrow(cmps))){
  tmp <- pks[pks$target_compound_id == cmps$compound_id[i],]
  cmps$RT[i] <- mean(tmp$rtmed)
  cmps$tissues[i] <- paste(unique(tmp$target_ion_adduct[
    grep("tissues", tmp$dataset)]), collapse = "; ")
  cmps$maturation[i] <- paste(unique(tmp$target_ion_adduct[
    grep("maturation", tmp$dataset)]), collapse = "; ")
}
cmps$class <- factor(cmps$class, levels = cmps_levels)
cmps <- cmps[order(-cmps$instd, cmps$class, cmps$C, cmps$db, cmps$RT),]
cmps$cmp <- paste0("C", formatC(seq(nrow(cmps)), width = nchar(nrow(cmps)), 
                                flag = "0"))
sum(cmps$instd == 0)
for(s in c("tissues", "maturation")){
  print(s)
  print(paste("Total compounds:", sum(cmps[,s] != "")))
  print(paste("Identified compounds:", 
              sum(cmps[cmps$class != "unk", s] != "")))
  print(paste("Unknown compounds:", sum(cmps[cmps$class == "unk", s] != "")))
  print("###########################################")
}
cmps$RT <- sprintf("%.2f", cmps$RT)

# TGs >3 possibilities
tg <- read.csv("data/compounds.csv")
tg <- tg[grep("TG", tg$compound),]
tg <- tg[grep("TG", tg$comments),]
for(i in seq(nrow(tg))){
  cmps$name[which(cmps$name == tg$compound[i])] <- tg$comments[i]
}

DT::datatable(
  cmps[,c("cmp", "name", "formula", "RT", "tissues", "maturation")],
  rownames = FALSE,
  options = list(pageLength = nrow(cmps))
)
rm(i, tmp)
write.csv(cmps[,c("cmp", "name", "formula", "RT", "tissues", "maturation")], 
          "output/Table_3.csv", row.names = FALSE)
```


# Figures

## Venn Diagram

```{r venn}
for(s in c("tissues", "maturation")){
  for(p in c("POS", "NEG")){
    pks <- get(paste("pks", s, p, sep = "_"))
    pks <- pks[!is.na(pks$score_rt) & !grepl("-", pks$target_compound_id),]
    pks <- pks[!grepl("tail", pks$target_name),]
    assign(paste("pks", s, p, sep = "_"), pks)
  }}

tissues <- unique(
  c(pks_tissues_NEG$target_name, pks_tissues_POS$target_name))
maturat <- unique(
  c(pks_maturation_NEG$target_name, pks_maturation_POS$target_name))
tissues <- tissues[!is.na(tissues)]
maturat <- maturat[!is.na(maturat)]
tissues <- tissues[!tissues %in% cmps$name[cmps$instd==1]]
maturat <- maturat[!maturat %in% cmps$name[cmps$instd==1]]

VennDiagram::draw.pairwise.venn(
  area1 = length(tissues),
  area2 = length(maturat),
  cross.area = length(tissues[tissues %in% maturat]),
  category = c("Tissues", "Ripening")
)
```


## Classes of compounds

```{r classes, eval = TRUE}
sum(cmps$class[cmps$instd == 0] != "nonFA")  # Total of annotated lipids
length(unique(
  cmps$class[cmps$instd == 0 & cmps$class != "nonFA"])) # Classes of lipids

cmps$class_super <- NA
cmps$class_super[which(
  cmps$class %in% c("LPA", "LPC", "LPE", "LPG", "LPI", "LPS",
                    "PA", "mPA", "dmPA", "PC", "PE", "PI", "PG", 
                    "PS"))] <- "(lyso)PL"
cmps$class_super[which(cmps$class %in% c(
  "DGMG", "MGDG", "DGDG", "SQDG", "acMGDG", "MG", "DG", "TG", "TG;O"))] <- "GL"
cmps$class_super[which(cmps$class %in% c(
  "SM", "Cer", "Cer;O3", "Cer;O4", 
  "HexCer", "HexCer;O3", "HexCer;O4", "LactCer"))] <- "SP"
cmps$class_super[cmps$class %in% c("CAR", "FFA", "ST", "Glc-ST", 
                                   "others", "nonFA")] <- "others"
unique(cmps$class[is.na(cmps$class_super)])
cmps$class <- as.character(cmps$class)
webr::PieDonut(
  cmps[cmps$instd == 0 & cmps$class != "nonFA",], 
  ggplot2::aes(pies = class_super, donuts = class), 
  showPieName = FALSE)

cmps$class2 <- cmps$class
cmps$class2[cmps$class %in% c(
  "DGMG", "MGDG", "DGDG", "SQDG", "acMGDG")] <- "others"
cmps$class2[cmps$class %in% c("LPE", "PE")] <- "(L)PE"
cmps$class2[cmps$class %in% c("PA", "mPA", "dmPA")] <- "(dm/m)PA"
cmps$class2[cmps$class %in% c("LPC", "PC")] <- "(L)PC"
cmps$class2[cmps$class %in% c("ST", "Glc-ST")] <- "(Glc)ST"
cmps$class2[cmps$class %in% c("Cer;O3", "HexCer;O3")] <- "(Hex)Cer;O3"
cmps$class2[cmps$class %in% c("Cer;O4", "HexCer;O4")] <- "(Hex)Cer;O4"
webr::PieDonut(
  cmps[cmps$instd == 0 & cmps$class2 != "nonFA",], 
  ggplot2::aes(pies = class_super, donuts = class2),
  labelposition = 1, showPieName = FALSE)

tmp <- data.frame(table(cmps$class[cmps$instd == 0 & cmps$class != "nonFA"]))
tmp$perc <- (tmp$Freq*100)/sum(tmp$Freq)
tmp <- tmp[order(tmp$Freq, decreasing = TRUE),]
head(tmp)
```


## Fatty acyl chains

```{r fa-chains}
idx <- which(cmps$instd == 0) # discard IS
cl <- unique(cmps$class[idx])
fa <- gsub(".*\\ ", "", cmps$name[idx]) # remove class from name
fa <- unlist(strsplit(fa, "_")) # separate FA chains
fa <- fa[grepl("[0-9]", substr(fa, 1, 1))]
fa <- fa[as.numeric(gsub(":.*", "", fa)) < 28]

tmp <- data.frame(table(fa))
tmp$perc <- (tmp$Freq*100)/sum(tmp$Freq)
tmp <- tmp[order(tmp$Freq, decreasing = TRUE),]
head(tmp)

C <- as.numeric(gsub(":.*", "", fa))
tmp <- data.frame(table(C))
tmp$perc <- (tmp$Freq*100)/sum(tmp$Freq)
tmp <- tmp[order(tmp$Freq, decreasing = TRUE),]
head(tmp)
range(C)
unique(C[C %% 2 != 0])[order(unique(C[C %% 2 != 0]))]

tmp <- cmps[cmps$instd == 0,]
prop.table(table(tmp$class[grep(paste(paste0(unique(C[C %% 2 != 0])[
  order(unique(C[C %% 2 != 0]))], ":"), collapse = "|"), tmp$name)]))*100

db <- as.numeric(gsub(".*:", "", fa))
tmp <- data.frame(table(db))
tmp$perc <- (tmp$Freq*100)/sum(tmp$Freq)
tmp <- tmp[order(tmp$Freq, decreasing = TRUE),]
head(tmp)
range(db)

fa <- unique(fa) # get unique values
fa <- fa[!is.na(fa)]
links <- data.frame(expand.grid(cl, fa))
colnames(links) <- c("source", "target")
links$value <- 0
for(i in seq(nrow(links))){
  links$value[i] <- length(grep(
    links$target[i], unlist(
      strsplit(cmps$name[cmps$class == links$source[i]], "_"))))
}
links <- links[links$value > 0,]
links <- links[order(links$source, as.character(links$target)),]
nodes <- data.frame(
  name = 
    c(levels(droplevels(links$source)), unique(as.character(
      links$target)[order(as.character(links$target))]))
  
)
links$IDsource <- match(links$source, nodes$name)-1 
links$IDtarget <- match(links$target, nodes$name)-1
networkD3::sankeyNetwork(Links = links, Nodes = nodes,
                         Source = "IDsource", Target = "IDtarget",
                         Value = "value", NodeID = "name", 
                         sinksRight = FALSE, iterations = 0)

links2 <- links[links$source %in% cmps$class[cmps$class_super == "GL"],]
nodes <- data.frame(
  name = c(levels(droplevels(links2$source)), unique(as.character(
    links2$target)[order(as.character(links2$target))]))
)
links2$IDsource <- match(links2$source, nodes$name)-1 
links2$IDtarget <- match(links2$target, nodes$name)-1
networkD3::sankeyNetwork(Links = links2, Nodes = nodes,
                         Source = "IDsource", Target = "IDtarget",
                         Value = "value", NodeID = "name", 
                         sinksRight = FALSE, iterations = 0)

links2 <- links[links$source %in% cmps$class[cmps$class_super == "(lyso)PL"],]
nodes <- data.frame(
  name = c(levels(droplevels(links2$source)), unique(as.character(
    links2$target)[order(as.character(links2$target))]))
)
links2$IDsource <- match(links2$source, nodes$name)-1 
links2$IDtarget <- match(links2$target, nodes$name)-1
networkD3::sankeyNetwork(Links = links2, Nodes = nodes,
                         Source = "IDsource", Target = "IDtarget",
                         Value = "value", NodeID = "name", 
                         sinksRight = FALSE, iterations = 0)


links$target <- as.character(links$target)
links$C <- as.numeric(gsub(":.*", "", links$target))
links$db <- as.numeric(gsub(".*:", "", links$target))

table(links$C, links$db)
links[links$target == "17:3",]

rm(cl, fa, links, i, nodes)
```


## KMD

```{r kmd}
ref <- c(12, 1, 16, 14, 31, 32)
names(ref) <- c("C", "H", "O", "N", "P", "S")
cmps$NM <- NA
for(i in seq(nrow(cmps))){
  if(cmps$formula[i] != "X"){
    myform <- countElements(cmps$formula[i])
    pippo <- ref[names(myform[[1]])]
    cmps$NM[i] <- t(pippo) %*% myform[[1]]
  } else {
    cmps$NM[i] <- floor(cmps$exactmass[i])
  }
}
cmps$KMD_H <- cmps$NM - (cmps$exactmass * ref["H"] / calculateMass("H"))
myform <- countElements("CH2")
pippo <- ref[names(myform[[1]])]
cmps$KMD_CH2 <- cmps$NM - (
  cmps$exactmass * as.numeric(
    t(pippo) %*% myform[[1]]) / calculateMass("CH2")
)


idx <- which(cmps$instd == 0) # discard IS
plot(cmps$KMD_CH2[idx], cmps$KMD_H[idx], xlab = "KMD(CH2)", ylab = "KMD(H)", 
     col = "white")
# indicate the C:db (for the unknowns indicate the name):
text(cmps$KMD_CH2, cmps$KMD_H, paste(cmps$C, cmps$db, sep = ":"), 
     col = col_cmps[cmps$class], cex = 0.8)

# define colors by number of double bonds:
mycols <- inlmisc::GetColors(max(cmps$db) + 1, "discrete rainbow")
names(mycols) <- c(seq(max(cmps$db)), 0)

# get the data from standards NOT included in study compounds:
stds <- compounds(CompDb(db_file), c(
  "compound_id", "name", "formula", "exactmass", "rtime", "class", 
  "std", "instd", "C", "db"))
stds <- stds[stds$std == 1 & stds$instd == 0, ]
stds <- stds[!stds$name %in% cmps$name, ]
stds <- stds[stds$rtime > 0, ]
if(nrow(stds) > 0){
  stds$exactmass <- NA
  stds$NM <- NA
  for(i in seq(nrow(stds))){
    stds$exactmass[i] <- calculateMass(stds$formula[i])
    myform <- countElements(stds$formula[i])
    pippo <- ref[names(myform[[1]])]
    stds$NM[i] <- t(pippo) %*% myform[[1]]
  }
  stds$KMD_H <- stds$NM - (stds$exactmass * ref["H"] / calculateMass("H"))
  stds$RT <- stds$rtime
}

cmps$class <- factor(cmps$class, levels = cmps_levels)
cl <- unique(cmps$class[cmps$instd == 0])
cl <- cl[!cl %in% c("others", "nonFA")]
pdf("output/FS1_KMD.pdf", paper = "a4", height = 4*3, width = 2*4)
par(mfrow = c(2, 1), mar = c(4, 3.5, 2, 0.5))
for(i in seq(length(cl))){
  # get the data from compounds of class "i":
  tmp <- rbind(
    cmps[cmps$class == cl[i], c("compound_id", "class", "name", "RT", 
                                "KMD_H", "std", "instd", "C", "db")],
    stds[stds$class == cl[i], c("compound_id", "class", "name", "RT", 
                                "KMD_H", "std", "instd", "C", "db")]
  )
  tmp <- tmp[tmp$instd == 0, ]
  if(nrow(tmp) > 1){
    # define ylims:
    if(cl[i] %in% c("TG")){
      y <- 0.1
    } else {
      y <- 0.05
    }
    # do the plot:
    plot(tmp$RT, tmp$KMD_H, col = "white", xlab = "", ylab = "", bty="l",
         ylim = c(min(tmp$KMD_H) - y, max(tmp$KMD_H) + 0.05))
    points(tmp$RT, tmp$KMD_H, pch = 19, cex = 2, 
           col = mycols[as.character(tmp$db)])
    idx <- which(tmp$std == 1) # compounds with STDs
    points(tmp$RT[idx], tmp$KMD_H[idx], pch = 3, cex = 2)
    idx <- which(tmp$name %in% cmps$name) # study compounds
    points(tmp$RT[idx], tmp$KMD_H[idx], pch = 1, cex = 3)
    abline(h=unique(round(tmp$KMD_H,2)), lty =2 , col = "grey")
    tmp$Cdb <- paste(tmp$C, tmp$db, sep = ":")
    if(any(duplicated(tmp$Cdb))){
      d <- tmp$Cdb[duplicated(tmp$Cdb)]
      x <- tmp$RT[!duplicated(tmp$Cdb)]
      y <- tmp$KMD_H[!duplicated(tmp$Cdb)]
      z <- tmp$Cdb[!duplicated(tmp$Cdb)]
      for(j in seq(length(d))){
        x[which(z == d[j])] <- mean(as.numeric(tmp$RT[tmp$Cdb == d[j]]))
      }
      text(x, y, z, pos = 1, offset = 1)
    } else {
      text(tmp$RT, tmp$KMD_H, tmp$Cdb, pos = 1, offset = 1)
    }
    title(cl[i], line = 0)
    title(xlab = "RT (min)", ylab = "KMD(H)", line = 2.5)
    # add diagonal lines:
    for(j in unique(tmp$db[duplicated(tmp$db)])){
      idx <- which(tmp$db == j)
      # only if there is >1 non-isomeric compound with "j" double bonds:
      if(length(unique(round(tmp$KMD_H[idx], 2))) > 1){
        if(!(cl[i] == "TG;O" & j > 3)){
          lines(as.numeric(tmp$RT[idx]), 
                predict(lm(tmp$KMD_H[idx]~as.numeric(tmp$RT[idx]))), lwd = 4, 
                col = paste0(mycols[as.character(j)], "60"))
        }
      }
    } # close db "j"
  } # close "if(nrow(tmp) > 0)"
} # close cl "i"
dev.off()
```


# Save output

```{r save}
pks <- rbind(pks_tissues_POS, pks_maturation_POS,
             pks_tissues_NEG, pks_maturation_NEG)
save(pks, db_file, cmps,
     dda_xdata_tissues_POS, dda_xdata_tissues_NEG,
     dda_xdata_maturation_POS, dda_xdata_maturation_NEG,
     dda_spectra_tissues_POS, dda_spectra_tissues_NEG,
     dda_spectra_maturation_POS, dda_spectra_maturation_NEG,
     ft_tissues_POS, ft_tissues_NEG,
     ft_maturation_POS, ft_maturation_NEG, 
     file = "output/identifications.RData")
```


# Session information

```{r session}
Sys.time()-startpoint
#devtools::session_info()
```
